created functional buttons for the author's sole use to perfom CRUD operations on their blog posts
