+++ app/controllers/author_controller.rb
@@ -1,0 +1,3 @@
+class AuthorController < ActionController::Base
+  protect_from_forgery with: :exception
+end
