+++ app/controllers/blog_controller.rb
@@ -1,2 +1,3 @@
 class BlogController < ActionController::Base
+  protect_from_forgery with: :exception
 end
