+++ app/views/author/posts/_post.html.erb
@@ -1,0 +1,6 @@
+<tr>
+  <td><%= link_to post.title, author_post_path(post) %></td>
+  <td></td>
+  <td><%= link_to 'Edit', edit_author_post_path(post), class: 'btn btn-sm btn-warning' %></td>
+  <td><%= link_to 'Remove', author_post_path(post), method: :delete, data: { confirm: 'Are you sure?' }, class: 'btn btn-sm btn-danger' %></td>
+</tr>
