+++ app/views/author/posts/_post.json.jbuilder
@@ -1,0 +1,2 @@
+json.extract! post, :id, :title, :body, :description, :slug, :created_at, :updated_at
+json.url post_url(post, format: :json)
