+++ app/views/author/posts/edit.html.erb
@@ -1,5 +1,5 @@
 <% provide(:page_title, "Edit #{@post.title}") %>
-
+<% provide(:author, 'active') %>
 
 <h1>Editing Post</h1>
