+++ app/views/author/posts/index.html.erb
@@ -1,0 +1,16 @@
+<% provide(:page_title, 'All Posts') %>
+<% provide(:author, 'active') %>
+<br>
+<div>
+   <%= link_to 'New post', new_author_post_path, class: 'btn btn-primary btn-lg' %>
+</div>
+<br>
+<div class="row">
+  <table class="table table-striped">
+    <tr>
+      <th scope="col">Title</th>
+    </tr>
+    <%= render @posts %>
+
+  </table>
+</div>
