+++ app/views/author/posts/index.json.jbuilder
@@ -1,0 +1,1 @@
+json.array! @posts, partial: 'posts/post', as: :post
