+++ app/views/author/posts/new.html.erb
@@ -1,4 +1,5 @@
 <% provide(:page_title, 'New Post') %>
+<% provide(:author, 'active') %>
 
 
 <h1>New Post</h1>
