+++ app/views/author/posts/show.html.erb
@@ -1,0 +1,16 @@
+<% provide(:page_title, @post.title) %>
+<% provide(:author, 'active') %>
+
+<br>
+<div>
+  <%= link_to 'Edit', edit_author_post_path(@post), class: 'btn btn-sm btn-warning' %>
+  <%= link_to 'Remove', author_post_path(@post), method: :delete, data: { confirm: 'Are you sure?' }, class: 'btn btn-sm btn-danger' %>
+</div>
+
+<div class="row">
+  <div class="col-md-10 offset-md-1 col-lg-8 offset-lg-2 text-xs-center">
+    <h1><%= @post.title %></h1>
+    <p><%= @post.display_day_published %></p>
+    <%= @post.body.html_safe %>
+  </div>
+</div>
