+++ app/views/author/posts/show.json.jbuilder
@@ -1,0 +1,1 @@
+json.partial! "posts/post", post: @post
