+++ app/views/layouts/_navbar.html.erb
@@ -8,6 +8,7 @@
         <a class="nav-link" href="#"><%= link_to 'Blog', root_path, class: "nav-link #{yield(:blog_active)}" %></a>
         <a class="nav-link" href="#"><%= link_to 'Media', root_path, class: "nav-link #{yield(:blog_active)}" %></a>
         <a class="nav-link" href="#"><%= link_to 'Wyncode/Wyntalent', wyncode_path, class: "nav-link #{yield(:wyncode_active)}" %>
+        <a class="nav-link" href="#"><%= link_to 'My Posts', author_posts_path, class: "nav-link #{yield(:author)}" %></a>
 </a>
     </div>
 </nav>
