+++ app/views/layouts/author.html.erb
@@ -10,10 +10,8 @@
   </head>
 
   <body>
-    <div class="">
-      <%= render 'layouts/navbar' %>
-    </div>
-    <div class="container <%=controller_name %> <%= action_name%>" >
+    <%= render 'layouts/navbar' %>
+    <div class="container <%=controller_name %> <%= action_name%> " >
       <%= yield %>
     </div>
   </body>
